+++ src/Resources/app/administration/src/module/swag-paypal/index.js
@@ -6,6 +6,7 @@ import './components/sw-paypal-express';
 import './components/sw-paypal-installment';
 import './components/sw-paypal-plus';
 import './components/sw-paypal-spb';
+import './components/sw-paypal-locale-field';
 
 import deDE from './snippet/de-DE.json';
 import enGB from './snippet/en-GB.json';